rename `text_modify` mod to `text_guard`.
