+++ src/interface/bin/mod.rs
@@ -1,6 +1,6 @@
 pub mod style;
 pub mod text_body;
-pub mod text_modify;
+pub mod text_guard;
 pub mod text_state;
 
 use std::any::Any;
