+++ src/interface/bin/text_guard.rs
